+++ app/src/main/java/com/interiordesign3d/di/Modules.kt
@@ -1,11 +1,13 @@
 package com.interiordesign3d.di
 
+import com.interiordesign3d.ads.AdsManager
 import com.interiordesign3d.ui.screen.catalogue.CatalogueViewModel
 import com.interiordesign3d.ui.screen.catalogue.item.CatalogueItemViewModel
 import com.interiordesign3d.ui.screen.designer.DesignerViewModel
 import com.interiordesign3d.ui.screen.home.HomeViewModel
 import com.interiordesign3d.ui.screen.project.ProjectViewModel
 import com.interiordesign3d.ui.screen.language.LanguageViewModel
+import com.interiordesign3d.ui.screen.language.setting.LanguageSettingViewModel
 import com.interiordesign3d.ui.screen.onboard.OnboardViewModel
 import com.interiordesign3d.ui.screen.select.SelectViewModel
 import com.interiordesign3d.ui.screen.settings.SettingsViewModel
@@ -14,6 +16,8 @@ import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
 val viewModelModule = module {
+    single { AdsManager() }
+
     viewModelOf(::HomeViewModel)
     viewModelOf(::ProjectViewModel)
     viewModelOf(::CatalogueViewModel)
@@ -22,6 +26,7 @@ val viewModelModule = module {
     viewModelOf(::SplashViewModel)
     viewModelOf(::OnboardViewModel)
     viewModelOf(::LanguageViewModel)
+    viewModelOf(::LanguageSettingViewModel)
     viewModelOf(::SelectViewModel)
     viewModelOf(::DesignerViewModel)
 }