+++ app/src/main/java/com/interiordesign3d/ui/screen/onboard/state/OnboardState.kt
@@ -1,9 +1,0 @@
-package com.interiordesign3d.ui.screen.onboard.state
-
-import androidx.compose.runtime.Stable
-import com.interiordesign3d.common.base.BaseScreenState
-
-@Stable
-open class OnboardState : BaseScreenState() {
-    open fun onFinish() {}
-}